fitbit_to_aws_s3/requirements.txt
python-dotenv==1.2.1
requests==2.31.0
fastapi==0.121.0
uvicorn==0.38.0
pandas==2.3.3
boto3==1.34.0